On stage 4 added true multy-threading
